refactor: placePinSource — the /map pin rendering facade (#1170) (#1198)
* refactor(map): extract the pin/cluster/heatmap rendering facade #1170

The /map page carried ~730 lines of pure rendering machinery: three
GeoJSON sources, fifteen layers, badge sprites, the Place-to-feature
build with lazy name resolution, boosted routing, the zoom-boundary
re-sync, and heatmap visibility juggling. createPlacePinSource owns all
of it behind install/render/resyncForZoom/heatmap calls, with the
page-owned context (saved ids, enriched cache, display lang, the
elementsLoaded + __mapPlacesCount hooks) injected as deps. The page
keeps what genuinely interacts with its state: click/hover handlers,
spiderfy, camera, label palette, and the render-signature gating that
decides WHEN to render — the facade only knows HOW.

render() keeps the #1194 contract: it renders exactly the list it is
given; all visibility policy stays upstream in selectVisiblePlaces.
Facade behavior is pinned by 8 unit tests (boost routing both sides of
the zoom boundary, flip-only re-sync, heatmap seed-on-enable and
skip-while-off, conceal/reveal zoom, persisted-state re-apply,
dep-injected feature properties).


* refactor(map): review follow-ups — dead import, pinned id-array contract #1170

Drop the GeoJSONSource type import the extraction left behind (nothing
in the page touches sources anymore — that was the point), type the
exported CUSTOM_SOURCE_IDS/CUSTOM_LAYER_IDS as readonly so the
applyBasemap carryover contract can't be mutated from outside, and use
HEATMAP_STORAGE_KEY in the test instead of a hardcoded literal.


---------
